More 12 bit changes + Moo 1.2 Support
MyTags.cs needed some more changes (divide by 4096 instead of 1024)
to fully support the 12 bit ADC. I also added code for the Moo 1.2
to support the new digital accelerometer. The digital and analog
accelerometer use the same GUI pieces in the front end, as I figured
we'd never have a moo with both a digital and analog accelerometer.
However, that can be changed later.